Little refactoring the data generation things
