+++ src/main/java/top/offsetmonkey538/offsetend/datagen/DataGeneration.java
@@ -1,0 +1,18 @@
+package top.offsetmonkey538.offsetend.datagen;
+
+import net.fabricmc.fabric.api.datagen.v1.DataGeneratorEntrypoint;
+import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator;
+import top.offsetmonkey538.offsetend.datagen.language.ModEnglishLanguageProvider;
+
+public class DataGeneration implements DataGeneratorEntrypoint {
+
+    @Override
+    public void onInitializeDataGenerator(FabricDataGenerator fabricDataGenerator) {
+        FabricDataGenerator.Pack pack = fabricDataGenerator.createPack();
+
+        pack.addProvider(ModModelProvider::new);
+        pack.addProvider(ModBlockTagProvider::new);
+        pack.addProvider(ModEnglishLanguageProvider::new);
+        pack.addProvider(ModBlockLootTableProvider::new);
+    }
+}
