+++ src/main/java/top/offsetmonkey538/offsetend/datagen/ModBlockTagProvider.java
@@ -1,0 +1,20 @@
+package top.offsetmonkey538.offsetend.datagen;
+
+import net.fabricmc.fabric.api.datagen.v1.FabricDataOutput;
+import net.fabricmc.fabric.api.datagen.v1.provider.FabricTagProvider;
+import net.minecraft.registry.RegistryWrapper;
+import net.minecraft.registry.tag.BlockTags;
+import top.offsetmonkey538.offsetend.registry.ModBlocks;
+
+import java.util.concurrent.CompletableFuture;
+
+public class ModBlockTagProvider extends FabricTagProvider.BlockTagProvider {
+    public ModBlockTagProvider(FabricDataOutput output, CompletableFuture<RegistryWrapper.WrapperLookup> registriesFuture) {
+        super(output, registriesFuture);
+    }
+
+    @Override
+    protected void configure(RegistryWrapper.WrapperLookup arg) {
+        getOrCreateTagBuilder(BlockTags.FLOWER_POTS).add(ModBlocks.POTTED_END_MUSHROOM);
+    }
+}
