+++ src/main/java/top/offsetmonkey538/offsetend/datagen/ModModelProvider.java
@@ -1,0 +1,23 @@
+package top.offsetmonkey538.offsetend.datagen;
+
+import net.fabricmc.fabric.api.datagen.v1.FabricDataOutput;
+import net.fabricmc.fabric.api.datagen.v1.provider.FabricModelProvider;
+import net.minecraft.data.client.BlockStateModelGenerator;
+import net.minecraft.data.client.ItemModelGenerator;
+import top.offsetmonkey538.offsetend.registry.ModBlocks;
+
+public class ModModelProvider extends FabricModelProvider {
+    public ModModelProvider(FabricDataOutput output) {
+        super(output);
+    }
+
+    @Override
+    public void generateBlockStateModels(BlockStateModelGenerator blockStateModelGenerator) {
+        blockStateModelGenerator.registerFlowerPotPlant(ModBlocks.END_MUSHROOM, ModBlocks.POTTED_END_MUSHROOM, BlockStateModelGenerator.TintType.NOT_TINTED);
+    }
+
+    @Override
+    public void generateItemModels(ItemModelGenerator itemModelGenerator) {
+
+    }
+}
