+++ src/main/java/top/offsetmonkey538/offsetend/datagen/language/ModEnglishLanguageProvider.java
@@ -1,0 +1,16 @@
+package top.offsetmonkey538.offsetend.datagen.language;
+
+import net.fabricmc.fabric.api.datagen.v1.FabricDataOutput;
+import net.fabricmc.fabric.api.datagen.v1.provider.FabricLanguageProvider;
+import top.offsetmonkey538.offsetend.registry.ModBlocks;
+
+public class ModEnglishLanguageProvider extends FabricLanguageProvider {
+    public ModEnglishLanguageProvider(FabricDataOutput dataOutput) {
+        super(dataOutput);
+    }
+
+    @Override
+    public void generateTranslations(TranslationBuilder translationBuilder) {
+        translationBuilder.add(ModBlocks.END_MUSHROOM, "End Mushroom");
+    }
+}
